requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
gunicorn
requests